UPDT

+++ 42cursus/01-printf/printf/src/ft_printfselect.c
@@ -6,7 +6,7 @@
 /*   By: jalves-c < jalves-c@student.42lisboa.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/07 17:49:08 by jalves-c          #+#    #+#             */
-/*   Updated: 2022/12/20 19:00:34 by jalves-c         ###   ########.fr       */
+/*   Updated: 2022/12/22 16:04:03 by jalves-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	ft_printfselect(char c, va_list args)
 	else if (c == 'u')
 		return (ft_printfun(va_arg(args, unsigned int)));
 	else if (c == 'x')
-		return (ft_printflx(va_arg(args, int)));
+		return (ft_printflx((unsigned long long)va_arg(args, unsigned int)));
 	else if (c == 'X')
 		return (ft_printfux(va_arg(args, int)));
 	else if (c == 'p')
